+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Unpin `@rehooks/local-storage` now that it's no longer broken. Refs UIORGS-344.
 * Add `Note` field to contact people summary list. Refs UIORGS-352.
+* Organizations filters consistency in `ui-organizations` and `ui-plugin-find-organization` . Refs UIORGS-373.
 
 ## [4.0.0](https://github.com/folio-org/ui-organizations/tree/v4.0.0) (2023-02-22)
 [Full Changelog](https://github.com/folio-org/ui-organizations/compare/v3.3.1...v4.0.0)